Kendall, Spearman's Rank and Pearson corelogram
An important decision when you want to do a correlation analysis is whether to use parametric or non parametric correlation techniques. In that case, pearson is so usefull and parametric way to analyze correlation between each features. However, you have to be sure that model variables have linear relationship and also have less outliers for Pearson, if not you should choose Spearman's rank or Kendall correlogram methods to analyze for feature selection..