OLS regression
Backward elimination is one of the most significant methods to select model features. Therefore, OLS regression statistic is in role. Example of some OLS results meaning:
R-squared:  percentage variation in dependent that is explained by independent variables
AIC/BIC: A lower AIC represents to a better model. BIC stands for Bayesian information criteria and is a variant of AIC.
Durbin-watson: A value between 1 to 2 is preferred